requirements.txt
Pillow>=8.0.1
numpy>=1.17.3
pandas>=0.25.3
